Added keycodes and changed case of mousebutton enum
The reason the mousebutton enum has been changed to UPPERCASE, and the reason the keycode enum is also SCREAMING_SNAKE_CASE is that otherwise there would be conflicting names. The delete and return key for example would create a problem if they were lowercase. Since those keywords are case sensitive we can get around that by using SCREAMING_SNAKE_CASE.